requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
pydantic
python-dotenv
celery[redis]
redis
httpx
